move cli/config pkg to client/config
