+++ cli/flags/common.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 
 	"github.com/Sirupsen/logrus"
-	cliconfig "github.com/docker/docker/cli/config"
+	"github.com/docker/docker/client/config"
 	"github.com/docker/docker/opts"
 	"github.com/docker/go-connections/tlsconfig"
 	"github.com/spf13/pflag"
@@ -49,7 +49,7 @@ func NewCommonOptions() *CommonOptions {
 // InstallFlags adds flags for the common options on the FlagSet
 func (commonOpts *CommonOptions) InstallFlags(flags *pflag.FlagSet) {
 	if dockerCertPath == "" {
-		dockerCertPath = cliconfig.Dir()
+		dockerCertPath = config.Dir()
 	}
 
 	flags.BoolVarP(&commonOpts.Debug, "debug", "D", false, "Enable debug mode")
