+++ client/config/config.go
@@ -7,7 +7,7 @@ import (
 	"path/filepath"
 
 	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/cli/config/configfile"
+	"github.com/docker/docker/client/config/configfile"
 	"github.com/docker/docker/pkg/homedir"
 )
