+++ client/config/credentials/native_store.go
@@ -4,7 +4,7 @@ import (
 	"github.com/docker/docker-credential-helpers/client"
 	"github.com/docker/docker-credential-helpers/credentials"
 	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/cli/config/configfile"
+	"github.com/docker/docker/client/config/configfile"
 )
 
 const (
