+++ client/config/credentials/native_store_test.go
